Remove SWD_Base from terminal.s
Previous versions of terminal.s required a change to the
forth kernel, because they used the ramallot macro to
allocate the SWD ring buffers. This version calls allot from
swd-init instead. The downside of this approach is that swd-init
has to be called after the directory pointer has been initialized.